Moved DownloadManager code across .cpp files
