+++ src/downloadmanagement.cpp
@@ -1,5 +1,7 @@
 #include "downloadmanagement.h"
 
+#include <QThread>
+
 ////////////////////////////////////////////////////////////////////////////////
 // DowloadState
 ////////////////////////////////////////////////////////////////////////////////
@@ -42,6 +44,30 @@ DownloadManager::DownloadManager(const Library* lib, kiwix::Downloader *download
     , mp_downloader(downloader)
 {
     restoreDownloads();
+}
+
+DownloadManager::~DownloadManager()
+{
+    if ( mp_downloadUpdaterThread )
+    {
+        QThread* t = mp_downloadUpdaterThread;
+        mp_downloadUpdaterThread = nullptr; // tell the thread to terminate
+        t->wait();
+    }
+}
+
+void DownloadManager::startDownloadUpdaterThread()
+{
+    // so that DownloadInfo can be copied across threads
+    qRegisterMetaType<DownloadInfo>("DownloadInfo");
+
+    mp_downloadUpdaterThread = QThread::create([=]() {
+       while ( mp_downloadUpdaterThread != nullptr ) {
+            updateDownloads();
+            QThread::msleep(1000);
+        }
+    });
+    mp_downloadUpdaterThread->start();
 }
 
 void DownloadManager::restoreDownloads()
